backend/requirements.txt
# Core FastAPI
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.20
starlette==0.37.2

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
PyJWT==2.10.1

# Validation
pydantic==2.12.5
email-validator==2.3.0

# Security & Encryption
cryptography==46.0.3

# 2FA & Passkeys
pyotp==2.9.0
qrcode==8.2
webauthn==2.7.0

# HTTP Client
httpx==0.28.1

# Utils
python-dotenv==1.2.1
bleach==6.3.0
pillow==12.0.0
